refactor: test 코드를 domain 패키지로 옮김
